chore: expose CellsAsQueryConditionSchema
